refactor: generalize hyphen-shape helpers for a second ambiguous shape
Vinta chose a two-entry brand-suffix late fix in the Chrome extension (公視+, 影劇館+) instead of the classifier, after six prompt families failed to beat the rules on bundle plans. This is the preparatory structural step with no behavior change, kept separate so the next commit is purely the new shape.

sliceSentence, CJK, and a generalized indexOfNthSymbol(text, symbol, ordinal) move from shapes/hyphen-shape.ts into shapes/base.ts so a second ambiguous shape can share them. AmbiguousShape.edits() now receives the whole settled candidate instead of (settled, index) because the coming shape needs the unspaced neighbour of its symbol to decide a second edit. The hyphen shape and its tests follow the new signatures.
